[Privatization] Move FinalizeClassesWithoutChildrenRector here
