+++ rules/dead-code/src/Rector/Plus/RemoveDeadZeroAndOneOperationRector.php
@@ -21,7 +21,6 @@ use Rector\Core\Rector\AbstractRector;
 use Rector\Core\Util\StaticInstanceOf;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
-use Webmozart\Assert\Assert;
 
 /**
  * @see https://3v4l.org/I0BGs
