+++ rules/privatization/tests/Rector/Class_/FinalizeClassesWithoutChildrenRector/Fixture/issue_1639.php.inc
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rector\SOLID\Tests\Rector\Class_\FinalizeClassesWithoutChildrenRector\Fixture;
+namespace Rector\Privatization\Tests\Rector\Class_\FinalizeClassesWithoutChildrenRector\Fixture;
 
 use Doctrine\ORM\Mapping as ORM;
 use Snowflake\Snowflake;
