+++ rules/privatization/tests/Rector/Class_/FinalizeClassesWithoutChildrenRector/Fixture/skip.php.inc
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rector\Privatization\Tests\Rector\Class_\FinalizeClassesWithoutChildrenRector\Fixture;
+
+abstract class Skip
+{
+}
