+++ rules/privatization/tests/Rector/Class_/FinalizeClassesWithoutChildrenRector/Fixture/skip_embedable.php.inc
@@ -1,0 +1,12 @@
+<?php
+
+namespace Rector\Privatization\Tests\Rector\Class_\FinalizeClassesWithoutChildrenRector\Fixture;
+
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * @ORM\Embeddable
+ */
+class SkipEmbedable
+{
+}
