--- rules/solid/tests/Rector/Class_/FinalizeClassesWithoutChildrenRector/Fixture/skip.php.inc
@@ -1,7 +1,0 @@
-<?php
-
-namespace Rector\SOLID\Tests\Rector\Class_\FinalizeClassesWithoutChildrenRector\Fixture;
-
-abstract class Skip
-{
-}
