--- rules/solid/tests/Rector/Class_/FinalizeClassesWithoutChildrenRector/Fixture/skip_embedable.php.inc
@@ -1,12 +1,0 @@
-<?php
-
-namespace Rector\SOLID\Tests\Rector\Class_\FinalizeClassesWithoutChildrenRector\Fixture;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Embeddable
- */
-class SkipEmbedable
-{
-}
